Move hook internals to a seperate module
