+++ pluggy/__init__.py
@@ -1,8 +1,9 @@
 __version__ = '0.6.0'
 
 __all__ = ["PluginManager", "PluginValidationError", "HookCallError",
-           "HookspecMarker", "HookimplMarker", "_Result"]
+           "HookspecMarker", "HookimplMarker"]
 
 from .manager import *
 from .manager import _formatdef
 from .callers import _multicall, HookCallError, _Result, _legacymulticall
+from .hooks import HookspecMarker, HookimplMarker, _HookRelay, _HookCaller, varnames
